test(package): consolidate packed consumer coverage
